more D_MPI routines

+++ src/modprecision.f90
@@ -32,7 +32,7 @@ integer, parameter :: SHR_KIND_R4 = real32  !selected_real_kind( 6) ! 4 byte rea
 integer, parameter :: SHR_KIND_IN = kind(1) ! native integer
 
 !! Module specific precisions
-integer, parameter :: field_r = real32  ! Precision for the most common fields u, v, w
+integer, parameter :: field_r = real64  ! Precision for the most common fields u, v, w
                                         ! And all other fields that do not have
                                         ! their own kind and need to interoperate
 integer, parameter :: pois_r  = real64  ! NOTE: only real64 works with FFTW
